requirements.txt
pandas
torch
tensorflow
streamlit
transformers
datasets
numpy
psycopg2-binary
SQLAlchemy
Pillow